// frontend/src/types/liquor.ts
export interface Liquor {
  id: number;
  name: string;
  brand: string;
  category: string;
  country: string;
  alcohol_percent: number;
  volume: number;
  image_url: string;
  source: string;
  current_price: number;
  original_price: number;
  product_url: string;
  product_code: string;
}

export interface GroupedLiquor {
  name: string;
  brand: string;
  category: string;
  country: string;
  alcohol_percent: number;
  volume: number;
  image_url: string;
  vendors: {
    source: string;
    current_price: number;
    original_price: number;
    product_url: string;
  }[];
  lowest_price: number;
}